11. Container With Most Water
two pointers. Beginning with the widest path(intuitively), if left is
shorter, then left plus one, as the if there is area larger, then it
must comes with higher left wall to compensate the "lost width";if the
left is higher, then the right pointer minus one. continue the process
till the end.

1. intuitively choose a best answer to begin
2. make a hypothesis that there is a better one, then constitute the
best one..

116 ms, 14.1MB